:recycle: (ReFactor: Interfaces.tsx 삭제, 함수 표현식으로 수정)
Closes #5
